ut_FileStream

+++ include/nw4r/ut/ut_FileStream.h
@@ -16,14 +16,20 @@ public:
     public:
         FilePosition() : mFileSize(0), mFileOffset(0) {}
 
-        u32 GetFileSize() const { return mFileSize; }
-        void SetFileSize(u32 fileSize) { mFileSize = fileSize; }
+        u32 GetFileSize() const {
+            return mFileSize;
+        }
+        void SetFileSize(unsigned long fileSize) {
+            mFileSize = fileSize;
+        }
 
-        u32 Tell() const { return mFileOffset; }
+        u32 Tell() const {
+            return mFileOffset;
+        }
 
-        u32 Skip(s32 offset);
-        u32 Append(s32 offset);
-        void Seek(s32 offset, u32 origin);
+        u32 Skip(long offset);
+        u32 Append(long offset);
+        void Seek(long offset, unsigned long origin);
 
     private:
         u32 mFileSize;   // at 0x0
@@ -35,9 +41,9 @@ public:
     virtual ~FileStream() {} // at 0xC
 
     virtual u32 GetSize() const = 0;                             // at 0x40
-    virtual void Seek(s32 offset, u32 origin);                   // at 0x44
+    virtual void Seek(long offset, unsigned long origin);        // at 0x44
     virtual void Cancel();                                       // at 0x48
-    virtual bool CancelAsync(AsyncCallback callback, void* arg); // at 0x4C
+    virtual bool CancelAsync(AsyncCallback callback, void *arg); // at 0x4C
     virtual bool CanSeek() const = 0;                            // at 0x50
     virtual bool CanCancel() const = 0;                          // at 0x54
     virtual u32 Tell() const = 0;                                // at 0x58
